add authorization to business service

+++ app/business/business_service.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/minheq/kedul_server_main/auth"
 	"github.com/minheq/kedul_server_main/errors"
 )
 
@@ -16,6 +17,19 @@ type Service struct {
 // NewService constructor for AuthService
 func NewService(store Store) Service {
 	return Service{store: store}
+}
+
+// GetBusinessByID ...
+func (s *Service) GetBusinessByID(ctx context.Context, id string) (*Business, error) {
+	const op = "business/service.CreateBusiness"
+
+	business, err := s.store.GetBusinessByID(ctx, id)
+
+	if err != nil {
+		return nil, errors.Wrap(op, err, "failed to get business by id")
+	}
+
+	return business, nil
 }
 
 // CreateBusiness creates business
@@ -44,7 +58,7 @@ func (s *Service) CreateBusiness(ctx context.Context, userID string, name string
 }
 
 // UpdateBusiness updates business
-func (s *Service) UpdateBusiness(ctx context.Context, id string, name string, profileImageID string) (*Business, error) {
+func (s *Service) UpdateBusiness(ctx context.Context, id string, name string, profileImageID string, currentUser *auth.User) (*Business, error) {
 	const op = "business/service.UpdateBusiness"
 
 	existingBusiness, err := s.store.GetBusinessByName(ctx, name)
@@ -67,6 +81,10 @@ func (s *Service) UpdateBusiness(ctx context.Context, id string, name string, pr
 		return nil, errors.NotFound(op)
 	}
 
+	if business.UserID != currentUser.ID {
+		return nil, errors.Unauthorized(op)
+	}
+
 	business.UpdatedAt = time.Now()
 	business.Name = name
 	business.ProfileImageID = profileImageID
@@ -81,7 +99,7 @@ func (s *Service) UpdateBusiness(ctx context.Context, id string, name string, pr
 }
 
 // DeleteBusiness updates business
-func (s *Service) DeleteBusiness(ctx context.Context, id string) (*Business, error) {
+func (s *Service) DeleteBusiness(ctx context.Context, id string, currentUser *auth.User) (*Business, error) {
 	const op = "business/service.DeleteBusiness"
 
 	business, err := s.store.GetBusinessByID(ctx, id)
@@ -92,6 +110,10 @@ func (s *Service) DeleteBusiness(ctx context.Context, id string) (*Business, err
 
 	if business == nil {
 		return nil, errors.NotFound(op)
+	}
+
+	if business.UserID != currentUser.ID {
+		return nil, errors.Unauthorized(op)
 	}
 
 	err = s.store.DeleteBusiness(ctx, business)
